Machine learning in Action
This book sets out to introduce people to important machine learning
algorithms. Tools and applications using these algorithms are introduced
to give the reader an idea of how they are used in practice today. A
wide selection of machine learning books is available, which discuss the
mathematics, but discuss little of how to program the algorithms. This
book aims to be a bridge from algorithms presented in matrix form to an
actual functioning program. With that in mind, please note that this
book is heavy on code and light on mathematics.